requirements.txt
asn1crypto==0.24.0
aspy.yaml==1.3.0
astroid==2.2.5
beautifulsoup4==4.8.0
bs4==0.0.1
cached-property==1.5.1
certifi==2019.9.11
cffi==1.12.3
cfgv==2.0.1
chardet==3.0.4
commonmark==0.9.0
cryptography==2.7
dictdiffer==0.8.0
future==0.17.1
goodreads==0.3.2
identify==1.4.7
idna==2.8
importlib-metadata==0.23
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
nodeenv==1.3.3
nose==1.3.7
notion==0.0.24
omdb==0.10.1
pep8==1.7.1
pre-commit==1.18.3
pycparser==2.19
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-slugify==3.0.4
python-telegram-bot==12.1.1
pytz==2019.2
pyyaml==5.1.2
rauth==0.7.3
requests==2.22.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.3
text-unidecode==1.3
toml==0.10.0
tornado==6.0.3
typed-ast==1.4.0 ; implementation_name == 'cpython'
tzlocal==2.0.0
urllib3==1.25.5
virtualenv==16.7.5
wrapt==1.11.2
xmltodict==0.12.0
zipp==0.6.0